Fog diffuses as you ascend the mountain

+++ areas/world.gd
@@ -44,7 +44,16 @@ onready var fog_defaults := {
 	"end":env.environment.fog_depth_end
 }
 
+var env_override := false
+
+var time := 0.0
+var TIME_READY := 0.5
+
 const FOG_TWEEN_TIME := 2.5
+const FOG_MIN_HEIGHT := 1000
+const FOG_MAX_HEIGHT := 3500
+onready var FOG_DISTANCE_MIN = fog_defaults.end
+onready var FOG_DISTANCE_MAX = fog_defaults.end*2.0
 
 func _enter_tree():
 	if !Global.valid_game_state and ResourceLoader.exists(Global.save_path):
@@ -66,7 +75,9 @@ func _ready():
 	update_active_chunks(player_last_postion, true)
 
 func _process(delta):
+	time += delta
 	var player_new_position = player.global_transform.origin
+	apply_fog(player_new_position.y)
 	if (player_last_postion - player_new_position).length_squared() >= MIN_SQDIST_UPDATE:
 		get_tree().call_group("distance_activated", "process_player_distance", player_new_position)
 		detect_enemies(delta)
@@ -188,6 +199,12 @@ func load_async(ch:Spatial):
 	call_deferred("add_dynamic_content", ch, scn.instance())
 	loading = null
 
+func apply_fog(height: float):
+	var factor := clamp((height - FOG_MIN_HEIGHT)/(FOG_MAX_HEIGHT - FOG_MIN_HEIGHT), 0, 1)
+	fog_defaults.end = lerp(FOG_DISTANCE_MIN, FOG_DISTANCE_MAX, factor)
+	if !env_override:
+		env.environment.fog_depth_end = fog_defaults.end
+
 func load_sync(chunk: Spatial):
 	if chunk.name in chunk_unload_waitlist:
 		var _x = chunk_unload_waitlist.erase(chunk.name)
@@ -287,6 +304,9 @@ func get_wind_audio():
 	return $audio_wind
 
 func set_sun_enabled(enabled:bool):
+	if time < TIME_READY:
+		sun.visible = enabled
+		return
 	sun_tween.remove_all()
 	sun_tween.interpolate_property(sun, "light_energy",
 		sun.light_energy,
@@ -302,6 +322,12 @@ func set_sun_enabled(enabled:bool):
 	sun_tween.start()
 
 func set_fog_override(fog: Color, begin: float, end:float):
+	env_override = true
+	if time < TIME_READY:
+		env.environment.fog_color = fog
+		env.environment.fog_depth_begin = begin
+		env.environment.fog_depth_end = end
+		return
 	env_tween.remove_all()
 	env_tween.interpolate_property(env.environment, "fog_color",
 		env.environment.fog_color, fog,
@@ -318,6 +344,7 @@ func set_fog_override(fog: Color, begin: float, end:float):
 	env_tween.start()
 
 func clear_fog_override():
+	env_override = false
 	env_tween.remove_all()
 	env_tween.interpolate_property(env.environment, "fog_color",
 		env.environment.fog_color, fog_defaults.color,
